requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-cors-headers==4.9.0
daphne==4.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
httpx==0.25.2
asgiref==3.9.1
gunicorn==25.3.0
uuid6==2025.0.1
requests==2.31.0
whitenoise==6.12.0
pycountry==26.2.16
